refactor: restructure user module with proper folder organization
