+++ src/modules/auth/decorators/role.decorator.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { UserRole } from '../../user/user.entity';
+import { UserRole } from '../../user/entities/user.entity';
 
 export const Roles = (...roles: UserRole[]) => SetMetadata('roles', roles);
