+++ src/modules/resume/controllers/resume.controller.ts
@@ -13,7 +13,7 @@ import { ResumeService } from '../services/resume.service';
 import { JwtAuthGuard } from '../../auth/guards/auth.guard';
 import { RoleGuard } from '../../auth/guards/role.guard';
 import { Roles } from '../../auth/decorators/role.decorator';
-import { UserRole } from '../../user/user.entity';
+import { UserRole } from '../../user/entities/user.entity';
 import { RequestWithUser } from '../../auth/interfaces/request.interface';
 import { Resume } from '../entities/resume.entity';
 import { ResumeDetail } from '../entities/resume-detail.entity';
