--- src/modules/resume/resume.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ResumeService } from './resume.service';
-
-describe('ResumeService', () => {
-  let service: ResumeService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ResumeService],
-    }).compile();
-
-    service = module.get<ResumeService>(ResumeService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
