+++ src/modules/user/entities/user.entity.ts
@@ -1,34 +1,34 @@
 import {
   Entity,
-  Column,
   PrimaryGeneratedColumn,
-  CreateDateColumn,
-  UpdateDateColumn,
+  Column,
   ManyToOne,
+  JoinColumn,
   ManyToMany,
   JoinTable,
+  BeforeInsert,
+  BeforeUpdate,
 } from 'typeorm';
-import { Role } from '../../role/entities/role.entity';
-import { Permission } from '../../permission/entities/permission.entity';
-import { UserRole } from '../enums/user-role.enum';
+import { Exclude } from 'class-transformer';
 import * as argon2 from 'argon2';
+import { Role } from '../../role/entities/role.entity';
 
-@Entity()
+export enum UserRole {
+  ADMIN = 'ADMIN',
+  USER = 'USER',
+}
+
+@Entity('users')
 export class User {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column({ unique: true })
+  @Column({ length: 100, unique: true })
   email: string;
 
-  @Column()
+  @Column({ length: 100 })
+  @Exclude()
   password: string;
-
-  @Column()
-  firstName: string;
-
-  @Column()
-  lastName: string;
 
   @Column({
     type: 'enum',
@@ -36,9 +36,6 @@ export class User {
     default: UserRole.USER,
   })
   userRole: UserRole;
-
-  @Column({ default: true })
-  isActive: boolean;
 
   @Column({ nullable: true })
   refreshToken: string;
@@ -49,38 +46,65 @@ export class User {
   @Column({ nullable: true })
   resetPasswordExpires: Date;
 
-  @Column({ nullable: true })
-  mainRoleId: number;
+  @Column({ length: 50, nullable: true })
+  firstName?: string;
 
-  @ManyToOne(() => Role, { nullable: true })
-  mainRole: Role;
+  @Column({ length: 50, nullable: true })
+  lastName?: string;
 
-  @ManyToMany(() => Role)
-  @JoinTable()
-  additionalRoles: Role[];
-
-  @ManyToMany(() => Permission)
-  @JoinTable()
-  permissions: Permission[];
+  @Column({ default: true })
+  isActive: boolean;
 
   @Column({ nullable: true })
-  lastLoginAt: Date;
+  mainRoleId?: number;
 
-  @CreateDateColumn()
+  @ManyToOne(() => Role, (role) => role.users)
+  @JoinColumn({ name: 'mainRoleId' })
+  mainRole?: Role;
+
+  @ManyToMany(() => Role, (role) => role.additionalUsers)
+  @JoinTable({
+    name: 'user_additional_roles',
+    joinColumn: { name: 'userId', referencedColumnName: 'id' },
+    inverseJoinColumn: { name: 'roleId', referencedColumnName: 'id' },
+  })
+  additionalRoles?: Role[];
+
+  @Column({ type: 'timestamp', nullable: true })
+  lastLoginAt?: Date;
+
+  @Column({ type: 'timestamp', default: () => 'CURRENT_TIMESTAMP' })
   createdAt: Date;
 
-  @UpdateDateColumn()
+  @Column({
+    type: 'timestamp',
+    default: () => 'CURRENT_TIMESTAMP',
+    onUpdate: 'CURRENT_TIMESTAMP',
+  })
   updatedAt: Date;
 
-  get fullName(): string {
-    return `${this.firstName} ${this.lastName}`.trim();
-  }
-
-  async hashPassword(): Promise<void> {
-    this.password = await argon2.hash(this.password);
+  @BeforeInsert()
+  @BeforeUpdate()
+  async hashPassword() {
+    if (this.password) {
+      try {
+        this.password = await argon2.hash(this.password);
+      } catch (error) {
+        console.error('Error hashing password:', error);
+      }
+    }
   }
 
   async validatePassword(password: string): Promise<boolean> {
-    return await argon2.verify(this.password, password);
+    try {
+      return await argon2.verify(this.password, password);
+    } catch (error) {
+      console.error('Error validating password:', error);
+      return false;
+    }
+  }
+
+  get fullName(): string {
+    return `${this.firstName || ''} ${this.lastName || ''}`.trim();
   }
 }
