+++ src/modules/user/repositories/user.repository.ts
@@ -1,5 +1,5 @@
 import { Repository } from 'typeorm';
-import { User } from './user.entity';
+import { User } from '../entities/user.entity';
 
 export class UserRepository extends Repository<User> {
   async findByEmail(email: string): Promise<User | null> {
