--- src/modules/user/user.entity.ts
@@ -1,110 +1,0 @@
-import {
-  Entity,
-  PrimaryGeneratedColumn,
-  Column,
-  ManyToOne,
-  JoinColumn,
-  ManyToMany,
-  JoinTable,
-  BeforeInsert,
-  BeforeUpdate,
-} from 'typeorm';
-import { Exclude } from 'class-transformer';
-import * as argon2 from 'argon2';
-import { Role } from '../role/entities/role.entity';
-
-export enum UserRole {
-  ADMIN = 'ADMIN',
-  USER = 'USER',
-}
-
-@Entity('users')
-export class User {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ length: 100, unique: true })
-  email: string;
-
-  @Column({ length: 100 })
-  @Exclude()
-  password: string;
-
-  @Column({
-    type: 'enum',
-    enum: UserRole,
-    default: UserRole.USER,
-  })
-  userRole: UserRole;
-
-  @Column({ nullable: true })
-  refreshToken: string;
-
-  @Column({ nullable: true })
-  resetPasswordToken: string;
-
-  @Column({ nullable: true })
-  resetPasswordExpires: Date;
-
-  @Column({ length: 50, nullable: true })
-  firstName?: string;
-
-  @Column({ length: 50, nullable: true })
-  lastName?: string;
-
-  @Column({ default: true })
-  isActive: boolean;
-
-  @Column({ nullable: true })
-  mainRoleId?: number;
-
-  @ManyToOne(() => Role, (role) => role.users)
-  @JoinColumn({ name: 'mainRoleId' })
-  mainRole?: Role;
-
-  @ManyToMany(() => Role, (role) => role.additionalUsers)
-  @JoinTable({
-    name: 'user_additional_roles',
-    joinColumn: { name: 'userId', referencedColumnName: 'id' },
-    inverseJoinColumn: { name: 'roleId', referencedColumnName: 'id' },
-  })
-  additionalRoles?: Role[];
-
-  @Column({ type: 'timestamp', nullable: true })
-  lastLoginAt?: Date;
-
-  @Column({ type: 'timestamp', default: () => 'CURRENT_TIMESTAMP' })
-  createdAt: Date;
-
-  @Column({
-    type: 'timestamp',
-    default: () => 'CURRENT_TIMESTAMP',
-    onUpdate: 'CURRENT_TIMESTAMP',
-  })
-  updatedAt: Date;
-
-  @BeforeInsert()
-  @BeforeUpdate()
-  async hashPassword() {
-    if (this.password) {
-      try {
-        this.password = await argon2.hash(this.password);
-      } catch (error) {
-        console.error('Error hashing password:', error);
-      }
-    }
-  }
-
-  async validatePassword(password: string): Promise<boolean> {
-    try {
-      return await argon2.verify(this.password, password);
-    } catch (error) {
-      console.error('Error validating password:', error);
-      return false;
-    }
-  }
-
-  get fullName(): string {
-    return `${this.firstName || ''} ${this.lastName || ''}`.trim();
-  }
-}
