+++ src/modules/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './user.entity';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
+import { User } from './entities/user.entity';
+import { UserService } from './services/user.service';
+import { UserController } from './controllers/user.controller';
 import { AuthModule } from '../auth/auth.module';
 
 @Module({
